extracted route refresh interface
