+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/directions/session/RouteRefresh.kt
@@ -1,0 +1,25 @@
+package com.mapbox.navigation.core.directions.session
+
+import com.mapbox.api.directions.v5.models.DirectionsRoute
+import com.mapbox.navigation.base.route.NavigationRoute
+import com.mapbox.navigation.base.route.NavigationRouterRefreshCallback
+
+internal interface RouteRefresh {
+    /**
+     * Refresh the traffic annotations for a given [DirectionsRoute]
+     *
+     * @param route DirectionsRoute the direction route to refresh
+     * @param legIndex Int the index of the current leg in the route
+     * @param callback Callback that gets notified with the results of the request
+     */
+    fun requestRouteRefresh(
+        route: NavigationRoute,
+        legIndex: Int,
+        callback: NavigationRouterRefreshCallback
+    ): Long
+
+    /**
+     * Cancels [requestRouteRefresh].
+     */
+    fun cancelRouteRefreshRequest(requestId: Long)
+}
